Added TCP socket implementation

+++ AsyncSocket.h
@@ -5,7 +5,7 @@
 class AsyncSocket
 {
 public:
-	virtual void connect(uint32_t ip, uint32_t port) = 0;
+	virtual void connect(uint32_t ip, uint16_t port) = 0;
 
 	virtual void send(const char* data, uint32_t length) = 0;
